LayoutTests/fast/dom/Window/watchos/no-websocket-in-watchos.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8">
    <title>No websocket in watchos</title>
    <script src="../../../../resources/testharness.js"></script>
    <script src="../../../../resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
test(() => {
    assert_equals(window.WebSocket, undefined);
}, 'Ensure that WebSocket is disabled in WatchOS');
    </script>
  </body>
</html>
